Adding all the sources that was missing because of an error in gitignore file.
At this point both the child firmware supports the ONLY_BEACON functionality. To set the clock, when the battery is plugged-in, the device turn on the receiver, and if it receives a CLIMB MASTER node broadcast packet with schedule wake up command, it set the wakeup clock and goes immediately in sleep mode.
On the master, by doing a long press on both button the device goes into the so called "children init" mode and broadcasts continuosly the schedule wake up command indipendently of the connection state.